decomp: synth follow-up — 6 FxSend TU recarves + 2 global levers, +139 (→13445)
Mic/FxSendDistortion recarve executed (unblocked by the mWav-mirage kill):
old pairings were correct-CLASS, wrong TU BOUNDARY. True Distortion360
start 0x82B32790; the 0x82B32F48-0x82B35FA0 region decoded as FIVE more
TUs and recarved: Delay (old span diffed the WRONG OBJ — synth/ vs
synth_xbox/ basename collision — and covered two TUs), Chorus (360-fns
live BEFORE its template cluster), Compress 360-part, EQ (real extent
0x82B343B8-0x82B34BC8), Flanger (new pin), Wah (extended to 0x82B35FA0).

Units: FxSendDistortion 26/26, FxSendEQ 27/27, FxSendChorus 25/25,
FxSendDelay 28/28, FxSendFlanger 13/13 COMPLETE; FxSendWah 27/28,
FxSendCompress 25/37. FlangerEffect::SetParameters 100 (Params
mFeedback/mWet are INT @0x10/0x14); Process rebuilt 54.3->77.9 banked
(old transliteration double-incremented a conflated loop counter —
real bug fixed; residue r30/r31+FPR mirror, permuter-class).

TWO GLOBAL LEVERS: (1) StandardEffect<T>::Reset() — the systematically
unpaired 8B thunk in every FxSend unit; one header line, 5 wins.
(2) OBJ_SET_TYPE_ENGINE — engine-era SetType inits static config BEFORE
the null-check, notify args PathName-before-ClassName; 5x SetType
64.3->100 (band3 keeps the old macro). Leads: CharLookAt::SetType 62.7,
DancerSequence::SetType 50.3 same disease. FxSend360 pure virtuals
(SyncEffectParams=0 etc.) bind overrides FxSend360-relative, fixing the
99.9 "mWav-flavored" mismatches. RB3-vs-DC3 layout deltas via Wii
oracle: FxSendEQ dropped mLRMode/mTransitionTime; FxSendWah gained
mDump/mUnk4C/mUnk28; Chorus/Flanger pct members are INT.

Zero losses (per-fn A/B; 10 entries moved between units during
recarves, all re-gained at 100). Leads: Synth.cpp recarve (12 fns
folded into Compress span incl GetHeadsetSubmix@Synth360); unpinned TU
0x82B34FE0-0x82B35604 (MeterEffect/Synapse family); fn_82B35F90 sized
delete MemFree(0x94,p) has no emitter in our objs.

+++ src/system/synth/FlangerEffect.h
@@ -9,8 +9,8 @@ public:
         float mDelayMs;  // delay time in milliseconds
         float mRate;     // LFO rate in Hz
         float mDepth;    // modulation depth (percentage)
-        float mFeedback; // feedback amount (percentage)
-        float mWet;      // wet/dry mix (percentage)
+        int mFeedback; // 0x10 feedback amount (percentage, int)
+        int mWet;      // 0x14 wet/dry mix (percentage, int)
     };
 
     ~FlangerEffect();
