+++ src/system/synth/FxSendChorus.h
@@ -31,10 +31,10 @@ protected:
     /** "Depth of pitch modulation, in centitones". Ranges from 0 to 1000. */
     float mDepth; // 0x68
     /** "Percent of output that is fed back to input". Ranges from 0 to 100. */
-    float mFeedbackPct; // 0x6c
+    int mFeedbackPct; // 0x6c
     /** "LFO phase offset between channels (for wider stereo effect)". Ranges from 0 to
      * 100. */
-    float mOffsetPct; // 0x70
+    int mOffsetPct; // 0x70
     /** "Sync modulation to song tempo?" */
     bool mTempoSync; // 0x74
     /** "Note value to synchronize modulation to". Sync type options: (sixteenth eighth
